Made scalar values permanently 64-bit on all platforms
Before it was dependent on the architecture being built for, which is frankly inappropriate for a relatively high level language.
Being able to rely on always the same size types across platforms is rather nice, and almost all platforms are natively 64-bit nowadays.
It will still build and run on 32-bit machines, it will just use 64-bit values there also.
Since this is only for Lobster values, the difference in size/speed is actually negliable at the moment.
Opt-in features for smaller scalar sizes on a per vector/struct basis may come later.