Fix for clang compile errors
https://stackoverflow.com/questions/13604137/definition-of-int64-t
Apparently Clang on Linux defines int64_t as long, but on OS X its long long ? Given that these types are not actually considered equal, that produces all sorts of fun errors and warnings when mixing different types that really "should" all be 64-bit signed. Who thinks of this stuff?