Fix crash building error message for function arity mismatch.
On linux, you can get a segmentation fault when you have a program
that passes the wrong number of parameters to a function.

The crash was occuring because Lex::Error was being passed an
invalid "ln" parameter from the Parser::ErrorAt() function.
ErrorAt() was taking the address of what->line when what
was null.