Don't reset gl_button frame counter on repeats
Most game code doesn't care about repeats. But they do care about when a button is pressed versus when it is held, and for how long. Repeats reset the counter, making it seem like a button was pressed when it wasn't. If a program actually wants repeats it can use gl_key_repeat or use the gl_button counter.