cubegen: Add remap_palettes argument to cg.load_vox
When loading a MagicaVoxel file, there are some files where the ordering
of the palette entries don't match the MagicaVoxel palette UI. In those
cases the file contains an "IMAP" chunk. However, this chunk just shows
how to remap the palette and voxel data so it matches the UI. If you
don't remap the palettes, everything will load and display correctly,
you just don't know which palette entries are which.

If you don't need this, it's better to set this argument to false, since
remapping the palette and voxel data takes some time.