Always pass through SDL_KEYUP messages, even if ImGui wants capture
When interacting with some ImGui elements, it is possible to get a
keydown event, then have ImGui capture keyboard input (via
IO.WantCaptureKeyboard) and subsequently never get the matching keyup
event. This is easy to reproduce by ctrl-clicking on a slider to edit
the text value directly.

Now we always pass through any keyup events. This doesn't really match
the suggested use of ImGui's IO.WantCaptureKeyboard, but on the other
hand it is hard to imagine why you'd ever want to send a keydown without
a matching keyup to the underlying input system.